Govspeak guide fixes
Fix a typo, move the 'answer summary' style to the bottom under a new 'deprecated styles' heading, so it's clear we don't use this anymore. Also, suffix the heading examples with 'example' so they don't get confused with the other names of elements.

Add 'govspeak' to the preview link so it's more descriptive link text.