Add name for CI workflow
Internal dependencies are not being auto-merged, despite adding the
configuration for GOV.UK Dependabot Merger (#442).

This is because GOV.UK Dependabot Merger looks for a workflow with the name
"CI" [2]. When name is absent, GitHub uses the workflow file path relative to
the root of the repository instead.

Add the name for the CI workflow.

[1]: #442
[2]: https://github.com/alphagov/govuk-dependabot-merger/blob/3206c703b2b4cf2dedef751313b3407485e89471/lib/pull_request.rb#L179