Move components that use ES6 to a separate file
This change is part of our preparation for upgrading to a version of the
govuk_publishing_components gem that includes the new GOV.UK Frontend V5.1

Since Grade X browsers don't support these components' ES6 JavaScript,
we want to prevent them evaluating it in the first place. In order to do
that, we load that JavaScript from a separate file and add the
type=”module” attribute to its script tag, which causes those browsers
to skip it.