Fix Sass mixed declarations deprecations
Sass is changing the way it works with "declarations mixed with nested
rules". We see the following deprecation warning when running
`rails assets:precompile` or in the server logs when loading a page that
uses affected Sass rules

```
Deprecation Warning: Sass's behavior for declarations that appear after nested
rules will be changing to match the behavior specified by CSS in an upcoming
version. To keep the existing behavior, move the declaration above the nested
rule. To opt into the new behavior, wrap the declaration in `& {}`.

More info: https://sass-lang.com/d/mixed-decls
```

The main thing we need to do is move a bunch of `@include`s below other
declarations within rules, similar to what was done in
govuk_publishing_components here: alphagov/govuk_publishing_components@e65096e

I've visually compared the site before and after updating and
recompiling the stylesheets and hard refreshing the page, checking that
components/elements using the affected rules look the same, and will
include screenshots in the associated pull request