Use govuk_app_config's Puma config
The govuk_app_config gem defines a bog-standard Puma config that most
(all?) of our other Rails apps use instead of defining their own.

I guess that the shared config hasn't been adopted by this app until now
because this app was originally deployed on Heroku and use of the
govuk_app_config gem was overlooked when that changed..?

The differences between the outgoing config and the shared one are:

  - use of `pidfile`
  - `RAILS_MAX_THREADS` default value

I assume the use a pid file doesn't make sense on our infrastructure,
otherwise it'd be used in the shared config.

Based on the commit history for this file, it looks like Govspeak
Preview moved from RAILS_MAX_THREADS defaulting to 5, to defaulting to
3 as part of a Rails app:update. The shared config has stuck to 5.