Provide nice error to user on failed conversion
We occasionally get `Encoding::CompatibilityError`s being reported in
Sentry. They all seem to be coming from
`GoogleDocsToGovspeak#to_govspeak` via the reverse_markdown gem, and the
error message reads `incompatible character encodings: UTF-8 and
ISO-8859-1 (Encoding::CompatibilityError)`

The form for converting a Google Doc to Govspeak mentions that you need
to remove images in the process. I tried submitting the form with a
Google Doc containing an image and got this error, so perhaps some if
not all of these errors stem from this kind of bad user journey?

Currently these exceptions are not handled, so we're just throwing a
generic 500 error page right at the user

This instead rescues the exception and provides a proper GOV.UK error
banner message asking the user to check the instructions and try again.
This does mean that the exceptions are effectively swallowed, but my
feeling is that this is a reasonable first step and users can submit
support requests if they are genuinely seeing errors after submitting
good data