Import styles for skip-link component
The styles for the skip-link component are currently imported in the layout-header component.

This import will be removed in a future release of the govuk_publishing_components gem to ensure the layout-header components only imports the styles it needs.

Adding this change in advance of the update to the gem will ensure the skip-link always renders correctly.