Upgrade govuk_publishing_components to 63.0.0
This upgrades govuk_publishing_components without downgrading
govuk_app_config

Just running `bundle update govuk_publishing_components` from main would
downgrade govuk_app_config. It wasn't super clear why, but the last
version of govuk_app_config pinned connection_pool to < 3 and we had it
on 3.0.2 here, so it's possible it was just picking govuk_app_config to
downgrade instead of its connection_pool dependency since it was higher
up the tree?

We downgraded connection_pool quite intentionally[[1][]], so I think the
preference would be to stick to govuk_app_config's latest version. So I
deleted the Gemfile.lock to get rid of the connection_pool 3.0.2 record
and ran `bundle install`, of which this is the result

[1]: https://github.com/alphagov/govuk_app_config/commit/f2995e14a81b1590f97fe3b412f151b41cc25354